[NAJIB&MIKE] edit update function and delete functions and add validation to return responseEntity with status code

+++ src/main/java/com/tw/apistackbase/controller/CompanyController.java
@@ -3,9 +3,11 @@ package com.tw.apistackbase.controller;
 import com.tw.apistackbase.core.Company;
 import com.tw.apistackbase.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/companies")
@@ -24,19 +26,26 @@ public class CompanyController {
         return repository.findOneByName(name);
     }
 
-    @PutMapping(produces = {"application/json"})
-    public Company updateCompanyInfo(@RequestBody Company company) {
-        Company company1 = repository.getOne(company.getId());
-        company1.setName(company.getName());
-        company1.setEmployees(company.getEmployees());
-        company1.setProfile(company.getProfile());
-        return repository.save(company1);
+    @PatchMapping(value = "/{id}", produces = {"application/json"})
+    public ResponseEntity<Company> updateCompanyInfo(@PathVariable Long id, @RequestBody Company company) {
+        Optional<Company> fetchedCompany = repository.findById(id);
+        if(fetchedCompany.isPresent()){
+            Company modifiedCompany = fetchedCompany.get();
+            modifiedCompany.setName(company.getName());
+            Company savedCompany = repository.save(modifiedCompany);
+            return new ResponseEntity<>(savedCompany, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping( value = "/{id}", produces = {"application/json"})
-    public List<Company> deleteCompany(@PathVariable Long id) {
-        repository.deleteById(id);
-        return repository.findAll();
+    public ResponseEntity<Company> deleteCompany(@PathVariable Long id) {
+        Optional<Company> fetchedCompany = repository.findById(id);
+        if(fetchedCompany.isPresent()){
+            repository.deleteById(id);
+            return new ResponseEntity<>(fetchedCompany.get(), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
